requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
python-docx==0.8.11
openpyxl==3.1.2
Werkzeug==2.3.7
Pillow==10.1.0
python-dateutil==2.8.2
email-validator==2.1.0
phonenumbers==8.13.22